Adjustments due to double controllers/repositories
